requirements.txt
Flask>=3.0,<4.0
gunicorn>=23.0,<24.0
azure-cosmos>=4.7,<5.0
